Fix DataGridPagerBase::getLast() fatal TypeError — floor() returns float under strict_types (#346)

+++ src/Infrastructure/Adapter/In/Web/DataGrid/Layout/DataGridPagerBase.php
@@ -290,7 +290,7 @@ abstract class DataGridPagerBase implements DataGridPagerInterface
     {
         return (($this->totalRows % $this->limitCount) === 0)
             ? $this->totalRows - $this->limitCount
-            : floor($this->totalRows / $this->limitCount) * $this->limitCount;
+            : (int)(floor($this->totalRows / $this->limitCount) * $this->limitCount);
     }
 
     /**
